+++ skbio/io/format/embl.py
@@ -1,0 +1,18 @@
+"""
+EMBL format (:mod:`skbio.io.format.embl`)
+===============================================
+
+.. currentmodule:: skbio.io.format.embl
+"""
+
+# ----------------------------------------------------------------------------
+# Copyright (c) 2013--, scikit-bio development team.
+#
+# Distributed under the terms of the Modified BSD License.
+#
+# The full license is in the file COPYING.txt, distributed with this software.
+# ----------------------------------------------------------------------------
+
+from skbio.io import create_format, EMBLFormatError
+
+embl = create_format('embl')